feat: 删除网络类型时判断有无网段和资产引用

+++ asset-web/src/main/java/com/antiy/asset/service/impl/AssetNettypeManageServiceImpl.java
@@ -1,12 +1,16 @@
 package com.antiy.asset.service.impl;
 
+import com.alibaba.fastjson.JSONObject;
 import com.antiy.asset.dao.AssetNettypeManageDao;
 import com.antiy.asset.entity.AssetNettypeManage;
 import com.antiy.asset.service.IAssetNettypeManageService;
+import com.antiy.asset.service.IAssetService;
+import com.antiy.asset.util.BaseClient;
 import com.antiy.asset.vo.query.AssetNettypeManageQuery;
 import com.antiy.asset.vo.request.AssetNettypeManageRequest;
 import com.antiy.asset.vo.response.AssetNettypeManageResponse;
 import com.antiy.biz.util.RedisUtil;
+import com.antiy.common.base.ActionResponse;
 import com.antiy.common.base.BaseConverter;
 import com.antiy.common.base.BaseRequest;
 import com.antiy.common.base.BaseServiceImpl;
@@ -19,6 +23,8 @@ import com.antiy.common.utils.ParamterExceptionUtils;
 import org.apache.commons.lang.ObjectUtils;
 import org.slf4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.core.ParameterizedTypeReference;
 import org.springframework.stereotype.Service;
 
 import javax.annotation.PostConstruct;
@@ -48,6 +54,13 @@ public class AssetNettypeManageServiceImpl extends BaseServiceImpl<AssetNettypeM
     private BaseConverter<AssetNettypeManage, AssetNettypeManageResponse> responseConverter;
     @Autowired
     private RedisUtil redisUtil;
+    @Autowired
+    private IAssetService assetService;
+    @Value("${segmentUrl}")
+    private String segmentUrl;
+
+    @Resource
+    private BaseClient baseClient;
 
     @PostConstruct
     public void init() throws Exception {
@@ -119,6 +132,18 @@ public class AssetNettypeManageServiceImpl extends BaseServiceImpl<AssetNettypeM
 
     @Override
     public String deleteAssetNettypeManageById(BaseRequest baseRequest) throws Exception {
+        JSONObject param = new JSONObject();
+        param.put("typeId", baseRequest.getStringId());
+        ActionResponse actionResponse = (ActionResponse)baseClient.post(param, new ParameterizedTypeReference<ActionResponse>() {
+        }, segmentUrl);
+        Integer body = (Integer) actionResponse.getBody();
+        if (body > 0) {
+            throw new BusinessException("该网络类型关联有网段，不能删除");
+        }
+        Integer assetCount = assetService.queryAssetCountByNetTypeId(baseRequest.getId());
+        if (assetCount > 0) {
+            throw new BusinessException("该网络类型关联有资产，不能删除");
+        }
         ParamterExceptionUtils.isBlank(baseRequest.getStringId(), "主键Id不能为空");
         String id = assetNettypeManageDao.deleteById(baseRequest.getStringId()).toString();
         redisUtil.hdel("net-type", id);
